Crate orml_xcm_support

XCM Support Module.

Overview

The XCM support module provides supporting traits, types and implementations, to support cross-chain message(XCM) integration with ORML modules.

Structs

IsNativeConcrete

A MatchesFungible implementation. It matches concrete fungible assets whose id could be converted into CurrencyId.

MultiCurrencyAdapter

The TransactAsset implementation, to handle MultiAsset deposit/withdraw. Note that teleport related functions are unimplemented.

MultiNativeAsset

A FilterAssetLocation implementation. Filters multi native assets whose reserve is same with origin.

Traits

UnknownAsset

Handlers unknown asset deposit and withdraw.